refactor : web과 terminal의 구조 분리
