+++ src/common/domain/Lotto.js
