+++ src/common/domain/LottoResult.js
