+++ src/common/domain/WinningLotto.js
