+++ src/common/lottoConstants/systemConstants.js
