+++ src/common/lottoConstants/systemMessage.js
