+++ src/common/util/formatNumber.js
