+++ src/common/validation/validateCondition.js
