+++ src/terminal/controller/PurchaseController.js
@@ -1,8 +1,8 @@
 import { getPrice } from "../service/InputService.js";
 import { getLottoArray, getLottoCount } from "../service/PurchaseService.js";
-import retryOnError from "../../util/retryOnError.js";
-import SYSTEM_MESSAGE from "../../lottoConstants/systemMessage.js";
 import OutputView from "../view/outputView.js";
+import retryOnError from "../../common/util/retryOnError.js";
+import SYSTEM_MESSAGE from "../../common/lottoConstants/systemMessage.js";
 
 export const PurchaseController = async () => {
   const price = await retryOnError(getPrice, OutputView.printError);
