+++ src/terminal/controller/ResultController.js
@@ -1,4 +1,4 @@
-import LottoResult from "../../domain/LottoResult.js";
+import LottoResult from "../../common/domain/LottoResult.js";
 
 export const ResultController = (winningLotto, lottoArray) => {
   const lottoResult = new LottoResult(winningLotto, lottoArray);
