+++ src/terminal/controller/WinningController.js
@@ -1,7 +1,7 @@
-import WinningLotto from "../../domain/WinningLotto.js";
 import { getBonusNumber, getWinningNumber } from "../service/InputService.js";
-import retryOnError from "../../util/retryOnError.js";
 import OutputView from "../view/outputView.js";
+import WinningLotto from "../../common/domain/WinningLotto.js";
+import retryOnError from "../../common/util/retryOnError.js";
 
 export const WinningController = async () => {
   const winningNumbers = await retryOnError(getWinningNumber, OutputView.printError);
