--- src/util/appendToParent.js
@@ -1,3 +1,0 @@
-export const appendToParent = (parentElement, ...childElement) => {
-  childElement.forEach((element) => document.querySelector(`${parentElement}`).appendChild(element));
-};
