+++ src/web/handler/purchaseHandler.js
@@ -1,15 +1,13 @@
+import { disableElement } from "../util/buttonActions.js";
 import { PurchaseController } from "../../terminal/controller/PurchaseController.js";
-import { disableElement } from "../../util/buttonActions.js";
 import { setState } from "../state/state.js";
-import { updateLottoArrayUI, updatelottoCountUI, updateResultButtonUI, updateWinningBonusUI } from "../ui/updateUI.js";
+import { updateUI } from "../ui/updateUI.js";
 
 export const purchaseLotto = async () => {
   const { lottoArray, lottoCount } = await PurchaseController();
+
   setState({ lottoArray: lottoArray, lottoCount: lottoCount });
-  updatelottoCountUI();
-  updateLottoArrayUI();
-  updateWinningBonusUI();
-  updateResultButtonUI();
+  updateUI();
 
   disableElement("purchase");
   disableElement("price");
