+++ src/web/handler/resultHandler.js
@@ -2,7 +2,7 @@ import { ResultController } from "../../terminal/controller/ResultController.js"
 import { WinningController } from "../../terminal/controller/WinningController.js";
 import { calculateProfitRate } from "../../terminal/service/ProfitService.js";
 import { getState, setState } from "../state/state.js";
-import { updateResultUI } from "../ui/updateUI.js";
+import { updateUI } from "../ui/updateUI.js";
 
 export const resultHandler = async () => {
   const { lottoCount, lottoArray } = getState();
@@ -10,7 +10,7 @@ export const resultHandler = async () => {
   const winningLotto = await WinningController();
   const matchingCount = ResultController(winningLotto, lottoArray);
   const profitRate = calculateProfitRate(matchingCount, lottoCount);
-  setState({ matchingCount: matchingCount, profitRate: profitRate });
 
-  updateResultUI();
+  setState({ matchingCount: matchingCount, profitRate: profitRate, isResultModalShow: true });
+  updateUI();
 };
