+++ src/web/ui/resetUI.js
@@ -1,4 +1,4 @@
-import { clearElement, removeElement } from "../../common/util/elementManager.js";
+import { clearElement, removeElement } from "../util/elementManager.js";
 
 export const resetUI = () => {
   const elementsToClear = [".count-prompt", ".lotto-numbers-container", ".winning-prompt", ".winning-bonus-container", ".result-button-container"];
