+++ src/web/util/elementManager.js
@@ -1,0 +1,13 @@
+export const appendElement = (parentElement, ...childElement) => {
+  childElement.forEach((element) => document.querySelector(`${parentElement}`).appendChild(element));
+};
+
+export const removeElement = (selector) => {
+  const element = document.querySelector(selector);
+  if (element) element.remove();
+};
+
+export const clearElement = (selector) => {
+  const element = document.querySelector(selector);
+  if (element) element.innerHTML = "";
+};
